Check if the domain is correctly set on windows VMIs
