+++ pkg/util/net/dns/dns_suite_test.go
@@ -1,0 +1,13 @@
+package dns_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestDns(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Dns Suite")
+}
